fix: extra Tab needed to navigate inside calendar
You can use Arrow keys to navigate insede
calendar but to be able to do this
with poped up calendar you need the
calendar to be in focus. To get calendar
in focus you need extra Tab press. This is
inconvenient so we make a popped up
calendar to be navigatable via keyboard
when popup's trigger is in focus.