fix: remove on focus outline from poped up picker
When interacting with poped up picker via mouse
it gets focused and therefore outlined. Outline
in this case only distracts user (popup's trigger
input field has outline) so we should
remove outline on focus from poped up picker.
Inline picker still outlined on focus
because this is the only way for user to understand
that the picker is in focus.